code/drv_mqtt/requirements.txt
system_logger_tool>=0.0.3
paho-mqtt>=1.5.0